requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-docx==0.8.11
